requirements.txt
absl-py==0.9.0
actionlib==1.12.0
aiohttp==3.6.2
angles==1.9.12
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
atari-py==0.2.6
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.2.1
bondpy==1.8.5
cachetools==4.1.1
camera-calibration==1.15.0
camera-calibration-parsers==1.11.13
catkin==0.7.26
catkin-pkg==0.4.22
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
colorama==0.4.3
controller-manager==0.18.1
controller-manager-msgs==0.18.1
cv-bridge==1.13.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
distro==1.5.0
dm-tree==0.1.5
docutils==0.16
dynamic-reconfigure==1.6.3
entrypoints==0.3
filelock==3.0.12
future==0.18.2
gast==0.3.3
gazebo-plugins==2.8.7
gazebo-ros==2.8.7
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.12
google==2.0.3
google-auth==1.19.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
gym==0.17.2
h5py==2.10.0
idna==2.9
idna-ssl==1.1.0
image-geometry==1.13.0
importlib-metadata==1.6.1
interactive-markers==1.11.5
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.15.1
joint-state-publisher==1.12.15
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
kdl-parser-py==1.13.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
laser-geometry==1.6.5
lz4==3.1.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
message-filters==1.14.6
mistune==0.8.4
mpi4py==3.0.3
msgpack==1.0.0
multidict==4.7.6
nbclient==0.5.0
nbconvert==6.0.6
nbformat==5.0.7
nest-asyncio==1.4.1
notebook==6.1.4
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.2.0.34
opencv-python-headless==4.2.0.34
opt-einsum==3.2.1
packaging==20.4
pandas==1.0.4
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pkg-resources==0.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.12.2
ptyprocess==0.6.0
py-spy==0.3.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==2.8.1
pycparser==2.20
pyglet==1.5.0
Pygments==2.7.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-qt-binding==0.4.2
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
qt-dotgraph==0.4.1
qt-gui==0.4.1
qt-gui-cpp==0.4.1
qt-gui-py-common==0.4.1
qtconsole==4.7.7
QtPy==1.9.0
ray==0.8.6
redis==3.4.1
requests==2.24.0
requests-oauthlib==1.3.0
resource-retriever==1.12.6
rosbag==1.14.6
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.6
roslaunch==1.14.6
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.6
rosmake==1.14.9
rosmaster==1.14.6
rosmsg==1.14.6
rosnode==1.14.6
rosparam==1.14.6
rospkg==1.2.8
rospy==1.14.6
rosservice==1.14.6
rostest==1.14.6
rostopic==1.14.6
rosunit==1.14.9
roswtf==1.14.6
rqt-action==0.4.9
rqt-bag==0.4.12
rqt-bag-plugins==0.4.12
rqt-console==0.4.9
rqt-dep==0.4.9
rqt-graph==0.4.11
rqt-gui==0.5.1
rqt-gui-py==0.5.1
rqt-image-view==0.4.14
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.9
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.1
rqt-py-console==0.4.8
rqt-reconfigure==0.5.1
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.9
rqt-robot-steering==0.5.10
rqt-runtime-monitor==0.5.7
rqt-rviz==0.6.0
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.11
rqt-web==0.4.8
rsa==4.6
rviz==1.13.12
scipy==1.4.1
seaborn==0.9.1
Send2Trash==1.5.0
sensor-msgs==1.12.7
six==1.15.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.5
soupsieve==2.0.1
stable-baselines==2.10.0
stable-baselines3==0.10.0
tabulate==0.8.7
tensorboard==1.14.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-probability==0.7.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tf==1.12.1
tf-conversions==1.12.1
tf2-geometry-msgs==0.6.5
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
topic-tools==1.14.6
torch==1.4.0
tornado==6.0.4
tqdm==4.56.0
traitlets==4.3.3
typing-extensions==3.7.4.2
urdfdom-py==0.4.3
urllib3==1.25.9
-e git+https://github.com/DRCL-USC/Reinforcement-Learning-for-Quadruped-Robots.git@71cf19e9981446387f9a7dd1df39c5c0058a0ddb#egg=usc_learning
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xacro==1.13.5
yarl==1.4.2
zipp==3.1.0
